requirement.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==20.2.0
cached-property==1.5.2
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
isodate==0.6.0
lxml==4.5.2
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.10
zeep==3.4.0
